First attempt at a Bash alias to use the required Ruby in any dir
This needs solving because Devon could potentially be run from any application
repo, and our Ruby applications run on a variety of different language versions.

Not only that, we have at least 2 Ruby version managers in active use at
Redbubble: `rbenv` and `chruby`. So if we're using Ruby, we need to work around
both of them.